Corrected return type to float for getElement32.

+++ src/main/java/us/ihmc/euclid/tuple3D/interfaces/Tuple3DReadOnly.java
@@ -130,7 +130,7 @@ public interface Tuple3DReadOnly
     * @return the value of the component.
     * @throws IndexOutOfBoundsException if {@code index} &notin; [0, 2].
     */
-   default double getElement32(int index)
+   default float getElement32(int index)
    {
       switch (index)
       {
